+++ bindings/java/src/org/sleuthkit/datamodel/ResultSetHelper.java
@@ -112,7 +112,7 @@ class ResultSetHelper {
 	File file(ResultSet rs, FileSystem fs) throws SQLException {
 		File f = new File(db, rs.getLong("obj_id"), rs.getLong("fs_obj_id"), rs.getLong("meta_addr"), rs.getShort("attr_type"),
 				rs.getShort("attr_id"), rs.getString("name"), rs.getShort("dir_type"),
-				rs.getShort("meta_type"), rs.getByte("dir_flags"), rs.getByte("meta_flags"), rs.getLong("size"),
+				rs.getShort("meta_type"), rs.getShort("dir_flags"), rs.getShort("meta_flags"), rs.getLong("size"),
 				rs.getLong("ctime"), rs.getLong("crtime"), rs.getLong("atime"), rs.getLong("mtime"),
 				rs.getShort("mode"), rs.getInt("uid"), rs.getInt("gid"), rs.getByte("known"), rs.getString("parent_path"), rs.getString("md5"));
 		f.setFileSystem(fs);
@@ -130,7 +130,7 @@ class ResultSetHelper {
 	Directory directory(ResultSet rs, FileSystem fs, String name) throws SQLException {
 		Directory dir = new Directory(db, rs.getLong("obj_id"), rs.getLong("fs_obj_id"), rs.getLong("meta_addr"), rs.getShort("attr_type"),
 				rs.getShort("attr_id"), name, rs.getShort("dir_type"),
-				rs.getShort("meta_type"), rs.getByte("dir_flags"), rs.getByte("meta_flags"), rs.getLong("size"),
+				rs.getShort("meta_type"), rs.getShort("dir_flags"), rs.getShort("meta_flags"), rs.getLong("size"),
 				rs.getLong("ctime"), rs.getLong("crtime"), rs.getLong("atime"), rs.getLong("mtime"),
 				rs.getShort("mode"), rs.getInt("uid"), rs.getInt("gid"), rs.getByte("known"), rs.getString("parent_path"), rs.getString("md5"));
 		dir.setFileSystem(fs);
